wchkdsk: change name to wchkdsk
